Move affected tests in a new module.
