+++ scalyr_agent/tests/util/util_test.py
@@ -34,10 +34,8 @@ import threading
 import uuid
 from mock import patch, MagicMock
 import six
-from six.moves import zip
 
 import scalyr_agent.util as scalyr_util
-from scalyr_agent import date_parsing_utils
 
 from scalyr_agent.util import (
     JsonReadFileException,
@@ -174,192 +172,6 @@ class TestUtil(ScalyrTestCase):
         expected = 1438872056123456
         actual = scalyr_util.microseconds_since_epoch(dt)
         self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_datetime(self):
-        s = "2015-08-06T14:40:56.123456Z"
-        expected = datetime.datetime(2015, 8, 6, 14, 40, 56, 123456)
-        actual = scalyr_util.rfc3339_to_datetime(s)
-
-        self.assertEquals(expected, actual)
-
-        s = "2015-08-06T14:40:56Z"
-        actual = scalyr_util.rfc3339_to_datetime(s)
-
-        self.assertEquals(datetime.datetime(2015, 8, 6, 14, 40, 56), actual)
-
-        s = "2015-08-06T14:40:56.123456"
-        actual = scalyr_util.rfc3339_to_datetime(s)
-        self.assertEquals(datetime.datetime(2015, 8, 6, 14, 40, 56), actual)
-
-    def test_rfc3339_to_datetime_with_and_without_strptime(self):
-        # Verify corectness between two different implementations of this function
-        input_strs = [
-            "2015-08-06T14:40:56.123456Z",
-            "2015-08-06T14:40:56Z",
-            "2015-08-06T14:40:56.123456",
-        ]
-        expected_dts = [
-            datetime.datetime(2015, 8, 6, 14, 40, 56, 123456),
-            datetime.datetime(2015, 8, 6, 14, 40, 56),
-            datetime.datetime(2015, 8, 6, 14, 40, 56),
-        ]
-
-        for input_str, expected_dt in zip(input_strs, expected_dts):
-            result_with_strptime = date_parsing_utils._rfc3339_to_datetime_strptime(
-                input_str
-            )
-            result_without_strptime = scalyr_util.rfc3339_to_datetime(input_str)
-
-            self.assertEqual(result_with_strptime, expected_dt)
-            self.assertEqual(result_with_strptime, result_without_strptime)
-
-    def test_rfc3339_to_datetime_truncated_nano(self):
-        s = "2015-08-06T14:40:56.123456789Z"
-        expected = datetime.datetime(2015, 8, 6, 14, 40, 56, 123456)
-        actual = scalyr_util.rfc3339_to_datetime(s)
-
-        self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_datetime_bad_format_date_and_time_separator(self):
-        s = "2015-08-06 14:40:56.123456789Z"
-        expected = None
-        actual = scalyr_util.rfc3339_to_datetime(s)
-        self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_datetime_bad_format_has_timezone(self):
-        # currently this function only handles UTC.  Remove this test if
-        # updated to be more flexible
-        s = "2015-08-06T14:40:56.123456789+04:00"
-        expected = None
-        actual = scalyr_util.rfc3339_to_datetime(s)
-        self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_nanoseconds_since_epoch(self):
-        s = "2015-08-06T14:40:56.123456Z"
-        expected = (
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56, 123456)
-            )
-            * 1000
-        )
-        actual = scalyr_util.rfc3339_to_nanoseconds_since_epoch(s)
-        self.assertEquals(expected, actual)
-
-        s = "2015-08-06T14:40:56.123456"
-        expected = (
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56)
-            )
-            * 1000
-        )
-        actual = scalyr_util.rfc3339_to_nanoseconds_since_epoch(s)
-        self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_nanoseconds_since_epoch_with_and_without_strptime(self):
-        # Verify corectness between two different implementations of this function
-        input_strs = [
-            "2015-08-06T14:40:56.123456Z",
-            "2015-08-06T14:40:56Z",
-            "2015-08-06T14:40:56.123456789Z",
-        ]
-        expected_tss = [
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56, 123456)
-            )
-            * 1000,
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56)
-            )
-            * 1000,
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56, 123456)
-            )
-            * 1000
-            + 789,
-        ]
-
-        for input_str, expected_ts in zip(input_strs, expected_tss):
-            result_with_strptime = date_parsing_utils._rfc3339_to_nanoseconds_since_epoch_strptime(
-                input_str
-            )
-            result_without_strptime = scalyr_util.rfc3339_to_nanoseconds_since_epoch(
-                input_str
-            )
-
-            self.assertEqual(result_with_strptime, expected_ts)
-            self.assertEqual(result_with_strptime, result_without_strptime)
-
-    def test_rfc3339_to_nanoseconds_since_epoch_no_fractions(self):
-        s = "2015-08-06T14:40:56"
-        expected = (
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56)
-            )
-            * 1000
-        )
-        actual = scalyr_util.rfc3339_to_nanoseconds_since_epoch(s)
-        self.assertEquals(expected, actual)
-
-        s = "2015-08-06T14:40:56Z"
-        expected = (
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56)
-            )
-            * 1000
-        )
-        actual = scalyr_util.rfc3339_to_nanoseconds_since_epoch(s)
-        self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_nanoseconds_since_epoch_some_fractions(self):
-        s = "2015-08-06T14:40:56.123Z"
-        expected = (
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56, 123000)
-            )
-            * 1000
-        )
-        actual = scalyr_util.rfc3339_to_nanoseconds_since_epoch(s)
-        self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_nanoseconds_since_epoch_many_fractions(self):
-        s = "2015-08-06T14:40:56.123456789Z"
-        expected = (
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56, 123456)
-            )
-            * 1000
-            + 789
-        )
-        actual = scalyr_util.rfc3339_to_nanoseconds_since_epoch(s)
-        self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_nanoseconds_since_epoch_too_many_fractions(self):
-        s = "2015-08-06T14:40:56.123456789999Z"
-        expected = (
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2015, 8, 6, 14, 40, 56, 123456)
-            )
-            * 1000
-            + 789
-        )
-        actual = scalyr_util.rfc3339_to_nanoseconds_since_epoch(s)
-        self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_nanoseconds_since_epoch_strange_value(self):
-        s = "2017-09-20T20:44:00.123456Z"
-        expected = (
-            scalyr_util.microseconds_since_epoch(
-                datetime.datetime(2017, 9, 20, 20, 44, 00, 123456)
-            )
-            * 1000
-        )
-        actual = scalyr_util.rfc3339_to_nanoseconds_since_epoch(s)
-        self.assertEquals(expected, actual)
-
-    def test_rfc3339_to_nanoseconds_since_epoch_bad_format_has_timezone(self):
-        s = "2015-08-06T14:40:56.123456789+04:00"
-        actual = scalyr_util.rfc3339_to_nanoseconds_since_epoch(s)
-        self.assertIs(None, actual)
 
     def test_uuid(self):
         first = scalyr_util.create_unique_id()
